feat.新增连续下载功能 refactor.重构代码结构，提高维护性和可阅读性 fix.修复转码结束后因缓冲区未清空导致的直接退出问题
