+++ .gitignore
@@ -16,7 +16,7 @@
 # Test binary, built with `go test -c`
 *.test
 
-# Output of the go coverage tool, specifically when used with LiteIDE
+# Output of the go coverage toolkit, specifically when used with LiteIDE
 *.out
 
 # Dependency directories (remove the comment below to include it)
