+++ internal/sso/sso.go
@@ -1,7 +1,7 @@
 package sso
 
 import (
-	"Bilibili_Downloader/internal/tool"
+	"Bilibili_Downloader/internal/toolkit"
 	"Bilibili_Downloader/pkg/cookie"
 	"Bilibili_Downloader/pkg/httpclient"
 	"encoding/json"
@@ -118,7 +118,7 @@ func HandleQRCodeLogin() error {
 		case 86090: // 已扫描未确认
 			fmt.Println("二维码已扫描，等待确认")
 		case 0: // 登录成功
-			tool.ClearScreen()
+			toolkit.ClearScreen()
 			fmt.Println("登录成功")
 			log.Println("登录成功")
 			cookie.StoreCookies(cookies)
