+++ internal/toolkit/download_info_struct.go
@@ -1,4 +1,4 @@
-package internal
+package toolkit
 
 type SegmentBase struct {
 	Initialization string `json:"Initialization"`
