+++ pkg/cookie/cookie.go
@@ -1,7 +1,7 @@
 package cookie
 
 import (
-	"Bilibili_Downloader/internal/tool"
+	"Bilibili_Downloader/internal/toolkit"
 	"encoding/json"
 	"fmt"
 	"log"
@@ -11,7 +11,7 @@ import (
 
 // 存储Cookie
 func StoreCookies(cookies []*http.Cookie) {
-	if err := tool.CheckAndCreateDir("./config"); err != nil {
+	if err := toolkit.CheckAndCreateDir("./config"); err != nil {
 		log.Println("视频输出目录检查或创建失败：", err)
 	}
 	// 创建文件存储 cookies
